requirements.txt
requests
beautifulsoup4
lxml
openpyxl
googlesearch-python
tld
